+++ bin/node/runtime/Cargo.toml
@@ -137,5 +137,6 @@ runtime-benchmarks = [
 	"pallet-timestamp/runtime-benchmarks",
 	"pallet-identity/runtime-benchmarks",
 	"pallet-balances/runtime-benchmarks",
+	"pallet-treasury/runtime-benchmarks",
 	"pallet-vesting/runtime-benchmarks",
 ]
